asc file experiments and new snp file generator
Various things to aid in analysing the asc file generated. This appears to be
working correctly.

Started new scan generator. Trying to do things a lot more robustly, because of
shifting requirements. On this have gotten a couple of placeholder functions,
but have (mostly) fully developed the "slew" function for use in higher-level
ones.