Various code fixes and tidying up.
Stopping for the night, in the meantime this is a commit to get things
a little bit more neat and tidy.
Added lots of docstrings (incomplete), tidied up a bit according to
PEP8.

Still need to add a bit of functionality, scan pattern is now done but
snp file write-out isn't done yet.