First possibly working version.
All functionality added. Still a few TODO items, but this file produces
a .snp output which should be executable by Field System.
Haven't bothered to use realistic values for the scan parameters yet, the
ones included in the file are just examples for visualising.

Still need quite a bit of PEP8 fixing.